reviewFunctions.js 추가
- 모델 검증 로직을 reviewFunctions.js로 이동
